moved main window UI code from main execution file
